% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eln4rmd.R
\name{elnjp}
\alias{elnjp}
\title{make new Japanese e-labnotebook}
\usage{
elnjp(add_name = FALSE, replace_date = FALSE)
}
\arguments{
\item{add_name}{If you want to add the name after the date as the file name,
add this argument.}

\item{replace_date}{If you want to use the specified date as the file name
instead of today's date, add this argument.}
}
\description{
make new Japanese e-labnotebook
}
