% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_eln.R
\name{new_els}
\alias{new_els}
\title{make new e-labnotebook}
\usage{
new_els(add_name = FALSE, replace_day = FALSE)
}
\arguments{
\item{add_name}{If you want to add the name after the date as the file name,
add this argument.}

\item{replace_day}{If you want to use the specified date as the file name
instead of today's date, add this argument.}
}
\description{
make new e-labnotebook
}
