% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eln4rmd.R
\name{render_elnjp_git}
\alias{render_elnjp_git}
\title{render function for Japanese e-labnotebook with GitHub}
\usage{
render_elnjp_git(Rmd_file)
}
\arguments{
\item{Rmd_file}{file name of R Markdown file}
}
\description{
render function for Japanese e-labnotebook with GitHub
}
