% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eln4rmd.R
\name{up_elnjp_osf}
\alias{up_elnjp_osf}
\title{upload Japanese e-labnotebook to OSF}
\usage{
up_elnjp_osf(add_name = FALSE, replace_date = FALSE, osf)
}
\arguments{
\item{add_name}{If you want to add the name after the date as the file name,
add this argument.}

\item{replace_date}{If you want to use the specified date as the file name
instead of today's date, add this argument.}

\item{osf}{URL of pdf directory in OSF}
}
\description{
upload Japanese e-labnotebook to OSF
}
