% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eln4rmd.R
\name{render_elnjp_osf}
\alias{render_elnjp_osf}
\title{render function for Japanese e-labnotebook with OSF}
\usage{
render_elnjp_osf(Rmd_file, osf)
}
\arguments{
\item{Rmd_file}{file name of R Markdown file}

\item{osf}{URL of pdf directory in OSF}
}
\description{
render function for Japanese e-labnotebook with OSF
}
